Update to 20240530

+++ sgi-csp-service/src/main/java/org/crue/hercules/sgi/csp/service/impl/ProyectoProrrogaServiceImpl.java
@@ -3,6 +3,7 @@ package org.crue.hercules.sgi.csp.service.impl;
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -11,11 +12,13 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.apache.commons.collections4.CollectionUtils;
 import org.crue.hercules.sgi.csp.exceptions.ProyectoNotFoundException;
 import org.crue.hercules.sgi.csp.exceptions.ProyectoProrrogaNotFoundException;
+import org.crue.hercules.sgi.csp.exceptions.ProyectoProrrogaWithRelatedProyectoFacturacionException;
 import org.crue.hercules.sgi.csp.model.Proyecto;
 import org.crue.hercules.sgi.csp.model.ProyectoEquipo;
 import org.crue.hercules.sgi.csp.model.ProyectoProrroga;
 import org.crue.hercules.sgi.csp.repository.ProrrogaDocumentoRepository;
 import org.crue.hercules.sgi.csp.repository.ProyectoEquipoRepository;
+import org.crue.hercules.sgi.csp.repository.ProyectoFacturacionRepository;
 import org.crue.hercules.sgi.csp.repository.ProyectoProrrogaRepository;
 import org.crue.hercules.sgi.csp.repository.ProyectoRepository;
 import org.crue.hercules.sgi.csp.repository.specification.ProyectoProrrogaSpecifications;
@@ -29,6 +32,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -36,6 +40,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Service
 @Slf4j
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
 
@@ -44,17 +49,7 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
   private final ProrrogaDocumentoRepository prorrogaDocumentoRepository;
   private final ProyectoEquipoRepository proyectoEquipoRepository;
   private final ProyectoHelper proyectoHelper;
-
-  public ProyectoProrrogaServiceImpl(ProyectoProrrogaRepository proyectoProrrogaRepository,
-      ProyectoRepository proyectoRepository, ProrrogaDocumentoRepository prorrogaDocumentoRepository,
-      ProyectoEquipoRepository proyectoEquipoRepository,
-      ProyectoHelper proyectoHelper) {
-    this.repository = proyectoProrrogaRepository;
-    this.proyectoRepository = proyectoRepository;
-    this.prorrogaDocumentoRepository = prorrogaDocumentoRepository;
-    this.proyectoEquipoRepository = proyectoEquipoRepository;
-    this.proyectoHelper = proyectoHelper;
-  }
+  private final ProyectoFacturacionRepository proyectoFacturacionRepository;
 
   /**
    * Guarda la entidad {@link ProyectoProrroga}.
@@ -77,7 +72,7 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
 
     // Actualizar nueva fecha de fin
     if (proyectoProrroga.getFechaFin() != null) {
-      this.actualizarFechaFin(returnValue);
+      this.actualizarFechaFin(returnValue.getProyectoId(), returnValue.getFechaFin());
     }
 
     // Se recalcula el número de prórroga en función de la ordenación de la fecha de
@@ -126,7 +121,7 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
 
       // Actualizar nueva fecha de fin
       if (actualizarFechaFin) {
-        this.actualizarFechaFin(returnValue);
+        this.actualizarFechaFin(returnValue.getProyectoId(), returnValue.getFechaFin());
       }
 
       log.debug("update(ProyectoProrroga ProyectoProrrogaActualizar) - end");
@@ -162,6 +157,10 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
       return proyectoProrroga;
     }).orElseThrow(() -> new ProyectoProrrogaNotFoundException(id));
 
+    if (proyectoFacturacionRepository.existsByProyectoProrrogaId(id)) {
+      throw new ProyectoProrrogaWithRelatedProyectoFacturacionException();
+    }
+
     // Borrado de los documentos asociados a la prórroga
     prorrogaDocumentoRepository.deleteByProyectoProrrogaId(id);
 
@@ -171,8 +170,18 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
     // concesión
     this.recalcularNumProrroga(proyectoProrrogaToDelete.getProyectoId());
 
-    log.debug("delete(Long id) - end");
+    if (proyectoProrrogaToDelete.getFechaFin() != null) {
+      Instant fechaFinNew = repository
+          .findAllByProyectoIdOrderByFechaConcesion(proyectoProrrogaToDelete.getProyectoId()).stream()
+          .map(ProyectoProrroga::getFechaFin)
+          .filter(fechaFin -> fechaFin != null)
+          .max(Comparator.naturalOrder())
+          .orElse(null);
 
+      this.actualizarFechaFin(proyectoProrrogaToDelete.getProyectoId(), fechaFinNew);
+    }
+
+    log.debug("delete(Long id) - end");
   }
 
   /**
@@ -236,16 +245,17 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
    * automáticamente la actualización de la fecha de fin al nuevo valor de fecha
    * de fin del proyecto.
    * 
-   * @param proyectoProrroga
+   * @param proyectoId  identificador del {@link Proyecto}
+   * @param fechaFinNew nueva fecha de fin
    */
-  private void actualizarFechaFin(ProyectoProrroga proyectoProrroga) {
-    log.debug("actualizarFechaFin(ProyectoProrroga proyectoProrroga) - start");
+  private void actualizarFechaFin(Long proyectoId, Instant fechaFinNew) {
+    log.debug("actualizarFechaFin(Long proyectoId, Instant fechaFinNew) - start");
 
-    Optional<Proyecto> proyecto = repository.getProyecto(proyectoProrroga.getId());
+    Optional<Proyecto> proyecto = proyectoRepository.findById(proyectoId);
 
     Assert.isTrue(
-        proyecto.isPresent() && proyectoProrroga.getFechaFin() != null
-            && proyectoProrroga.getFechaFin().compareTo(proyecto.get().getFechaInicio()) >= 0,
+        proyecto.isPresent() && ((fechaFinNew != null
+            && fechaFinNew.compareTo(proyecto.get().getFechaInicio()) >= 0) || fechaFinNew == null),
         "La fecha de fin debe ser posterior a la fecha de inicio del proyecto");
 
     // Se actualizan los miembros de equipo cuya fecha de fin coincida con la fecha
@@ -267,16 +277,17 @@ public class ProyectoProrrogaServiceImpl implements ProyectoProrrogaService {
     }
 
     if (CollectionUtils.isNotEmpty(miembros)) {
+      Instant fechaFinMiembrosNew = fechaFinNew != null ? fechaFinNew : fechaFin;
       miembros.stream().forEach(miembro -> {
-        miembro.setFechaFin(proyectoProrroga.getFechaFin());
+        miembro.setFechaFin(fechaFinMiembrosNew);
         proyectoEquipoRepository.save(miembro);
       });
     }
 
-    proyecto.get().setFechaFinDefinitiva(proyectoProrroga.getFechaFin());
+    proyecto.get().setFechaFinDefinitiva(fechaFinNew);
 
     proyectoRepository.save(proyecto.get());
-    log.debug("actualizarFechaFin(ProyectoProrroga proyectoProrroga) - end");
+    log.debug("actualizarFechaFin(Long proyectoId, Instant fechaFinNew) - end");
   }
 
   /**
